requirements.txt
django

django-cleanup

django-debug-toolbar

django-fixture-magic

django-grappelli

django-filebrowser

feedparser

future

psycopg2

psycopg2-binary

selenium

Pillow
